Refactor main function according to new folder structure and improve focus on display and game states
